Fix local test example in README
The local test example had invalid RSpec syntax. It also looks like
`Random.rand(3)` won't return `3` 33% of the time (the result gets
converted to an integer so it's always (0, 1, or 2 unless the random
number was _exactly_ 3.0) so it gets adjusted to be 1 to allow the
example to be more easily testable.